go/worker/compute/executor/committee: Move metrics to a new file
